requirements.txt
fastapi==0.109.0
pandas==2.1.4
uvicorn==0.25.0
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1
html5lib==1.1
sqlalchemy==2.0.27
pymysql==1.1.0
pydantic==2.6.1
numpy
python-dotenv